Add encode_url.js, s/parse_url.js, s/source.js
source.js contains the constructor for a Source type to streamline
adding new classes. At this point, there is significant refactoring yet
to do.

parse_url.js processes the query string, passing params to Source
objects to generate their URLs which are in turn used to create the
links.

encode_url.js simply html-escapes the input, then generates and appends
a link to the disambiguation page for that input. There is currently no
input validation.