Refactor parse_url.js and encode_url.js
Before, each link was set up individually, in a repetitive fashion. Not
DRY, to say the least. Plus it was all vanilla DOM manipulation, but
jquery is loaded and there's no performance reason not to use it.

So:

Process an array of link config objects, making it dead easy to add
or subtract.

Use jquery methods for all DOM manipulation