Extract [citation, source] types; detectType function; utils module
To further reduce parse_url.js to the core of the business logic and page
manipulation. Much shorter, much easier to follow now.

I debated also extracting formatUrl to the functions directory, so that
parse_url.js a really lean read, but though it's pure, it seemed appropriate to
keep the DOM manipulation together.