Add federal_rule subtypes to Westlaw and Lexis
Once it became apparent they needed to handle each type with its own function,
it was clear that source-by-source hacking was dumb. I replaced federal_rule in
the grand type_list with what had been the contents of subtypes; this
reorganization lead to dramatically simpler handling of subtypes in LII, as well.