Add live-server and `npm start`
* `npm install --save-dev live-server`
* Add "start" to "scripts" object:
  - `npm run watch-js & live-server --port=6789`

Previously, I developed using a home-baked non-caching WEBrick server
written in ruby. This is self-contained, provides a simpler interface
for kicking up a server (`npm start` instead of `npm run watch-js &
serve`), and auto-reloads on changes. It's also trivial to add in a
"watch-scss" script down the line if the custom styling starts making
more sense with preprocessing.