Snapshot after the last several months' work
RIO, the production version of this app hosted by Cornell Law's LII, is
hosted privately at LII's pleasure, which is as it should be. However,
as RIO has been developed under the MIT license throughout (per its
`package.json`); and as I am the sole author of the application's core
logic and architecture; and as it would be nice to be able to share the
level and style of code I chose to write given free rein; I'm
incorporating some of the work I did to expand and clean up this project
in its institutional home. I believe I have left in nothing specific to
LII, or indeed outside of the JS package itself.

Additionally, since this version will be seen by programmers rather than
lawyers, I added a quick "insert sample citation" feature to the link
builder.