Copy over what I can, ethically, from RIO
RIO, the production version of this app hosted by Cornell Law's LII, is
hosted privately at LII's pleasure, which is as it should be. That said:

* I am the sole author of the application's core logic and architecture
* RIO throughout has been developed under the MIT license (per its
  `package.json`), which allows free copying

So, while it's a shame to lose the rich git history of the RIO fork
(would that it were an actual github fork instead of a whole new history
beginning at a SLIGHTLY modified version of Lawfetcher's version as of
12/2015), it is legitimate to import the current state of the
application logic, purged of LII-specific content: so I am.